Sudoku: Make backtracking algorithm actually backtrack
Previously, if there were no viable solutions for the square, the entire
board would be deemed unsolveable. Now the algorithm actually
backtracks to a square with more than one possible value and contains
with an alternative value. This massively speeds up generation as we're
no longer search for a board where we guess the correct value first time
for every square