Sudoku: Reduce calls calculating possible values for a square
Previously, when backtracking we were calculating all the possible
values for a square at a given board state every time we wanted a new
one. Instead, we calculate this once and store them, then we either
interate though them in the backtracking algorithm, or randomly pick one
during board seeding.